+++ YACReaderLibrary/options_dialog.h
@@ -18,6 +18,7 @@ public:
 public slots:
     void editApiKey();
     void restoreOptions(QSettings *settings) override;
+    void saveOptions() override;
 
 private slots:
     void useBackgroundImageCheckClicked(bool checked);
@@ -35,6 +36,7 @@ private:
     QCheckBox *comicInfoXMLCheckbox;
     QSlider *recentIntervalSlider;
     QLabel *numDaysLabel;
+    QLineEdit *thirdPartyReaderEdit;
 
     // Libraries tab
     QCheckBox *updateLibrariesAtStartupCheck;